requirements.txt
pandas
ta
requests
python-dotenv
matplotlib
mplfinance
numpy
pybit
pandas